test(quic): hoist passive-rebind test body into a common suite
Per review feedback: rather than a third per-platform copy of the
passive-rebind test, extract the shared body into
QuicPassiveMigrationTestSuite (commonTest), mirroring QuicServerTestSuite
("guarantees test parity — same tests run on JVM, Linux, and Apple").

The only platform-specific pieces stay as abstract hooks:
  - testTlsConfig()       — cert resolution
  - createRebindingProxy() — the NAT-rebind UDP proxy (a new common
                             RebindingProxy interface)
  - wrapTestBody()         — UnsatisfiedLinkError → assumeTrue skip (JVM)

JVM QuicPassiveMigrationTests (#69) is rewritten as a thin subclass
providing the DatagramChannel proxy + native-lib skip; the new linuxX64
LinuxQuicPassiveMigrationTests provides the io_uring proxy. The single
common @Test now runs on every commonTest platform that supplies a
concrete subclass (JVM + Linux today; Apple can add one when its server
gap closes). JS has no subclass, so it doesn't run — correct, since JS
QUIC throws.

Android's copy stays in androidInstrumentedTest: that's a separate
on-device compilation that doesn't see commonTest (the reason it was a
copy to begin with) — noted in the suite KDoc.

Local: :socket-quic:linuxX64Test passive test runs (1.0s); JVM compiles
+ skips cleanly when the quiche native isn't staged (runs in CI);
ktlint clean.
